Description: zeroconfig patch was returning a select timeout of 0 for large timeouts due to unsigned value calling signed value.  This caused static dhcp leases (timeout 0xffffffff) to spin udhcpc.  Changed to unsigned for all timeouts.
git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/squeezeboxJive@456 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9

+++ src/system/busybox_1_7_stable/networking/udhcp/zeroconf.h
@@ -24,7 +24,7 @@ extern int zeroconf_fd;
 
 void zeroconf_init(int enable, uint8_t arp[], const char* interface, int ifindex);
 void zeroconf_event(int event);
-void zeroconf_set_smallest_timeout(struct timeval* tm, long int other_timeout);
+void zeroconf_set_smallest_timeout(struct timeval* tm, unsigned int other_timeout);
 
 /* Zeroconf constants, time constants in microseconds */
